requirements.txt
aiogram>=3.17.0
aiosqlite
apscheduler
python-dotenv
pytz
